pyproject.toml
[project]
name = "hotpotqa"
version = "0.1.0"
description = "HotpotQA Question Answering Project with IRCoT and Retrieval"
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    # Core dependencies
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    # NLP and text processing
    "nltk>=3.8.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "scikit-learn>=1.3.0",
    # Evaluation
    "ujson>=5.7.0",
    # Visualization
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "wordcloud>=1.9.0",
    # Notebooks
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    # Model inference
    "mistralai>=1.0.0",
    # Retrieval dependencies
    "faiss-cpu>=1.7.0",
    "sentence-transformers>=2.2.0",
    # Observability
    "langsmith>=0.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.1.0",
    "black>=23.3.0",
    "flake8>=6.0.0",
]
spacy = [
    "spacy>=3.5.0",
]

[project.scripts]
hotpotqa-preprocess = "scripts.run_preprocessing:main"
hotpotqa-validate = "scripts.validate_data:main"
hotpotqa-analyze = "scripts.analyze_dataset:main"

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
